% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertCharacter.R
\name{assertCharacter}
\alias{assertCharacter}
\title{Assert if an object is a character and fulfill certain conditions.}
\usage{
assertCharacter(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  unique = FALSE,
  named = FALSE,
  minNumCharacter = 0,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{unique}{Whether it has to contain unique elements.}

\item{named}{Whether it has to be named.}

\item{minNumCharacter}{Minimum number of characters.}

\item{call}{Call argument that will be passed to `cli`.}
}
\description{
Assert if an object is a character and fulfill certain conditions.
}
