% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertChoice.R
\name{assertChoice}
\alias{assertChoice}
\title{Assert if an object is a choice and fulfill certain conditions.}
\usage{
assertChoice(
  x,
  choices,
  length = NULL,
  na = FALSE,
  null = FALSE,
  unique = FALSE,
  named = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{choices}{Options that x can be.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{unique}{Whether it has to contain unique elements.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to `cli`.}
}
\description{
Assert if an object is a choice and fulfill certain conditions.
}
