% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/assertTibble.R
\name{assert_tibble}
\alias{assert_tibble}
\alias{assertTibble}
\title{Assert if an object is a tibble and fulfill certain conditions.}
\usage{
assert_tibble(
  x,
  numberColumns = NULL,
  numberRows = NULL,
  columns = NULL,
  null = FALSE,
  call = parent.frame()
)

assertTibble(
  x,
  numberColumns = NULL,
  numberRows = NULL,
  columns = NULL,
  null = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{numberColumns}{Number of columns.}

\item{numberRows}{Number of rows.}

\item{columns}{Name of columns that must be present.}

\item{null}{Whether it can be null.}

\item{call}{Call argument that will be passed to `cli`.}
}
\description{
Assert if an object is a tibble and fulfill certain conditions.
}
