% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardChecks.R
\name{checkCdm}
\alias{checkCdm}
\title{To check whether a cdm input is a valid cdm reference}
\usage{
checkCdm(cdm)
}
\arguments{
\item{cdm}{A cdm_reference created by CDMConnector}
}
\description{
To check whether a cdm input is a valid cdm reference
}
\examples{
\donttest{
library(DrugUtilisation)
cdm <- mockDrugUtilisation()
checkCdm(cdm)
}

}
