% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkInput}
\alias{checkInput}
\title{Function to check the inputs of a function}
\usage{
checkInput(..., .standard = TRUE)
}
\arguments{
\item{...}{Named parameters, the name of the parameter should be the name of
the test that you want to apply e.g. checkInput(name = parameters) it will
run the function checkName, and use parameter as name input}

\item{.standard}{Whether to include the standard tests included in the
package see the documentation to see the standard checks included}
}
\description{
Function to check the inputs of a function
}
