% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardChecks.R
\name{checkName}
\alias{checkName}
\title{To check whether a name input is a valid name}
\usage{
checkName(name, cdm)
}
\arguments{
\item{name}{Character string to check}

\item{cdm}{A cdm_reference created by CDMConnector}
}
\description{
To check whether a name input is a valid name
}
\examples{
\donttest{
library(DrugUtilisation)
cdm <- mockDrugUtilisation()
checkName("cohort3", cdm)
}

}
