% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputChecker-package.R
\docType{package}
\name{InputChecker-package}
\alias{InputChecker}
\alias{InputChecker-package}
\title{InputChecker: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
