% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{listInputCheck}
\alias{listInputCheck}
\title{List available inputs to check}
\usage{
listInputCheck(package)
}
\arguments{
\item{package}{Package to check}
}
\description{
List available inputs to check
}
