% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertClass.R
\name{assertClass}
\alias{assertClass}
\title{Assert that an object has a certain class.}
\usage{
assertClass(x, class, call = parent.frame())
}
\arguments{
\item{x}{To check.}

\item{class}{Expected class or classes.}

\item{call}{Call argument that will be passed to `cli`.}
}
\description{
Assert that an object has a certain class.
}
