% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkCdm}
\alias{checkCdm}
\title{Check the cdm}
\usage{
checkCdm(cdm, tablesToCheck = NULL)
}
\arguments{
\item{cdm}{Object to check if it is a valid cdm reference}
}
\value{
Informative error and warnings messages if the inputs don't pass the
designed check.
}
\description{
Check the cdm
}
\examples{
\donttest{
cdm <- 1
class(cdm) <- c("cdm_reference", class(cdm))
checkCdm(cdm)
}

}
