% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkName}
\alias{checkName}
\title{Check name for a new element in the cdm}
\usage{
checkName(name, cdm)
}
\arguments{
\item{name}{A character of length one with the name to be checked}

\item{cdm}{A cdm_reference}
}
\value{
Informative error and warnings messages if the inputs don't pass the
designed check.
}
\description{
Check name for a new element in the cdm
}
\examples{
\donttest{
cdm <- 1
class(cdm) <- c("cdm_reference", class(cdm))
checkCdm("new_element", cdm)
}

}
