% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputChecker-package.R
\docType{package}
\name{InputChecker-package}
\alias{InputChecker}
\alias{InputChecker-package}
\title{InputChecker: Chech Inputs for Functions in the DARWIN-EU Framewrok}
\description{
This package provides the necessary utilities for the DARWIN-EU packages, such as check environment to check inputs of functions consistently and create a mock cdm object.
}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
